Move secure_save() to ui-fileops
Fix group id getting.
Simplify a bit.
