+++ src/cache.cc
@@ -32,11 +32,9 @@
 
 #include <config.h>
 
-#include "intl.h"
 #include "main-defines.h"
 #include "md5-util.h"
 #include "options.h"
-#include "secure-save.h"
 #include "similar.h"
 #include "thumb-standard.h"
 #include "ui-fileops.h"
