+++ src/main.cc
@@ -79,7 +79,6 @@
 #include "metadata.h"
 #include "options.h"
 #include "pixbuf-util.h"
-#include "secure-save.h"
 #include "third-party/whereami.h"
 #include "thumb.h"
 #include "ui-fileops.h"
