+++ src/meson.build
@@ -147,8 +147,6 @@ main_sources = files('advanced-exif.cc',
 'search-and-run.h',
 'search.cc',
 'search.h',
-'secure-save.cc',
-'secure-save.h',
 'shortcuts.cc',
 'shortcuts.h',
 'similar.cc',
