+++ src/ui-fileops.h
@@ -102,5 +102,7 @@ guchar *map_file(const gchar *path, gsize &map_len);
 
 void pixbuf_gdk_known_extensions(GList **extensions_list);
 
+gboolean secure_save(const gchar *file_name, const gchar *contents, gsize length);
+
 #endif
 /* vim: set shiftwidth=8 softtabstop=0 cindent cinoptions={1s: */
